update split round, merge round

+++ controllers/request/RoundRequest.go
@@ -7,8 +7,8 @@ type AddRoundBody struct {
 
 type SplitRoundBody struct {
 	BookingUid string `json:"booking_uid"`
-	Hole       int64  `json:"hole"`
-	RoundIndex int64  `json:"round_index"`
+	Hole       int    `json:"hole"`
+	RoundIndex int    `json:"round_index"`
 }
 
 type MergeRoundBody struct {
